Restructured projects. Added Loggy.Api.DataAccess.MongoDb and Loggy.Api.Model.
Moved classes around, to decouple a little from Mongodb.
