+++ src/API/Loggy.Api.Model/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Loggy.Api.Model
+{
+	public class Class1
+	{
+	}
+}
