+++ src/API/Loggy.Api.Model/Model/LogEntry.cs
@@ -1,0 +1,7 @@
+﻿namespace Loggy.Api.Model.Model
+{
+	public class LogEntry
+	{
+		public string Name { get; set; }
+	}
+}
