+++ src/API/Loggy.Api.Model/Model/LogFieldDefinition.cs
@@ -1,0 +1,16 @@
+﻿namespace Loggy.Api.Model.Model
+{
+	public class LogFieldDefinition
+	{
+		public string FieldId { get; set; }
+		public string FieldName { get; set; }
+
+		public string DataTypeId { get; set; }
+
+		public LogFieldDefinition(string fieldId, string fieldName, string dataTypeId)
+		{
+			FieldId = fieldId;
+			FieldName = fieldName;
+		}
+	}
+}
