+++ src/API/Loggy.Api.Model/Model/LogFieldEntry.cs
@@ -1,0 +1,6 @@
+﻿namespace Loggy.Api.Model.Model
+{
+	public class LogFieldEntry
+	{
+	}
+}
