+++ src/API/Loggy.Api.Model/Model/User.cs
@@ -1,0 +1,14 @@
+﻿namespace Loggy.Api.Model.Model
+{
+	public class User
+	{
+		public string UserId { get; set; }
+		public string UserName { get; set; }
+
+		public User(string userId, string userName)
+		{
+			UserId = userId;
+			UserName = userName;
+		}
+	}
+}
