+++ src/API/Loggy.Api/GraphQlSchema/GraphQlQueryPostBody.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Newtonsoft.Json.Linq;
+﻿using Newtonsoft.Json.Linq;
 
-namespace Loggy.Api.Model
+namespace Loggy.Api.GraphQlSchema
 {
 	/// <summary>
 	/// Represents a graphQl query string, as POSTed to our GraphQlController.
