+++ src/API/Loggy.Api/Loggy.Api.csproj
@@ -18,4 +18,9 @@
     <PackageReference Include="MongoDB.Driver" Version="2.7.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\Loggy.Api.DataAccess.MongoDb\Loggy.Api.DataAccess.MongoDb.csproj" />
+    <ProjectReference Include="..\Loggy.Api.Model\Loggy.Api.Model.csproj" />
+  </ItemGroup>
+
 </Project>
