+++ src/LogChannels/Loggy.LogChannels.Api/Controllers/ValuesController.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Loggy.LogChannels.Api.Controllers
+namespace Loggy.Channels.ApiChannel.Controllers
 {
 	[Route("api/[controller]")]
 	[ApiController]
