+++ src/LogChannels/Loggy.LogChannels.Api/Program.cs
@@ -1,14 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore;
+﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
 
-namespace Loggy.LogChannels.Api
+namespace Loggy.Channels.ApiChannel
 {
 	public class Program
 	{
