+++ src/LogChannels/WebApplication1/Controllers/ValuesController.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
-namespace WebApplication1.Controllers
+namespace Loggy.Channels.EmailChannel.Controllers
 {
 	[Route("api/[controller]")]
 	[ApiController]
