+++ src/Loggy.Api.DataAccess.MongoDb/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Loggy.Api.DataAccess.MongoDb
+{
+	public class Class1
+	{
+	}
+}
