+++ src/Loggy.Api.DataAccess.MongoDb/DataAccess/LogSubjectRepository.cs
@@ -1,11 +1,8 @@
-﻿using System;
-
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using Loggy.Api.Model;
+using Loggy.Api.Model.Model;
 
-namespace Loggy.Api.DataAccess
+namespace Loggy.Api.DataAccess.MongoDb.DataAccess
 {
 	public interface ILogSubjectRepository
 	{
