+++ src/Loggy.Api.DataAccess.MongoDb/Loggy.Api.DataAccess.MongoDb.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Api\Loggy.Api.Model\Loggy.Api.Model.csproj" />
+  </ItemGroup>
+
+</Project>
