+++ src/Loggy.sln
@@ -21,6 +21,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Loggy.Channels.EmailChannel
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Loggy.Channels.ApiChannel", "LogChannels\Loggy.LogChannels.Api\Loggy.Channels.ApiChannel.csproj", "{258BD834-E609-4401-A23B-290E822833A7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Loggy.Api.Model", "Api\Loggy.Api.Model\Loggy.Api.Model.csproj", "{1BCF4089-DB76-42B7-91C5-A7590E2C635C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Loggy.Api.DataAccess.MongoDb", "Loggy.Api.DataAccess.MongoDb\Loggy.Api.DataAccess.MongoDb.csproj", "{77DD5941-C379-4376-A6FD-9C285BC5EFFC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,6 +51,14 @@ Global
 		{258BD834-E609-4401-A23B-290E822833A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{258BD834-E609-4401-A23B-290E822833A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{258BD834-E609-4401-A23B-290E822833A7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1BCF4089-DB76-42B7-91C5-A7590E2C635C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1BCF4089-DB76-42B7-91C5-A7590E2C635C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1BCF4089-DB76-42B7-91C5-A7590E2C635C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1BCF4089-DB76-42B7-91C5-A7590E2C635C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{77DD5941-C379-4376-A6FD-9C285BC5EFFC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{77DD5941-C379-4376-A6FD-9C285BC5EFFC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{77DD5941-C379-4376-A6FD-9C285BC5EFFC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{77DD5941-C379-4376-A6FD-9C285BC5EFFC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -57,6 +69,8 @@ Global
 		{A48B3776-DBD8-41CD-AD27-A045C9E0C65C} = {41EAC45B-2D7A-4B57-813A-7ABE8A1E65C6}
 		{10B87841-9CF6-4628-B873-01BB36C61446} = {870BD2EF-6999-4078-8867-1C4709FE4C9B}
 		{258BD834-E609-4401-A23B-290E822833A7} = {870BD2EF-6999-4078-8867-1C4709FE4C9B}
+		{1BCF4089-DB76-42B7-91C5-A7590E2C635C} = {71535CBB-F77E-4869-8667-5C8151996FEC}
+		{77DD5941-C379-4376-A6FD-9C285BC5EFFC} = {71535CBB-F77E-4869-8667-5C8151996FEC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {411BB96B-FA81-4DDA-A1CE-BFCEF55CBF4C}
